refactor: extract customer sign-up use case for hono
