+++ internal/services/src/use-cases/index.ts
@@ -1,2 +1,3 @@
 export { createPlan } from "./plan/create"
+export { signUp } from "./customer/sign-up"
 export { createSubscription } from "./subscription/create"
