+++ internal/trpc/src/router/lambda/planVersions/listByProjectUnprice.ts
@@ -3,7 +3,6 @@ import { getPlanVersionApiResponseSchema } from "@unprice/db/validators"
 import { z } from "zod"
 import { protectedProcedure } from "#trpc"
 
-
 // global endpoint, no need to check for feature access
 export const listByProjectUnprice = protectedProcedure
   .input(
