+++ internal/trpc/src/router/lambda/subscriptions/createPhase.ts
@@ -6,7 +6,6 @@ import {
 import { z } from "zod"
 import { protectedProjectProcedure } from "#trpc"
 
-
 export const createPhase = protectedProjectProcedure
   .input(subscriptionPhaseInsertSchema)
   .output(z.object({ phase: subscriptionPhaseSelectSchema }))
