+++ internal/trpc/src/router/lambda/subscriptions/updatePhase.ts
@@ -3,7 +3,6 @@ import { subscriptionPhaseSelectSchema } from "@unprice/db/validators"
 import { z } from "zod"
 import { protectedProjectProcedure } from "#trpc"
 
-
 export const updatePhase = protectedProjectProcedure
   .input(subscriptionPhaseSelectSchema)
   .output(z.object({ phase: subscriptionPhaseSelectSchema }))
